Moved context manager to optimizer.py
