+++ sdks/python/src/opik/api_objects/opik_client.py
@@ -10,10 +10,7 @@ from typing import (
     TypeVar,
     Union,
     Literal,
-    Type,
 )
-
-from types import TracebackType
 
 import httpx
 
@@ -53,83 +50,11 @@ from .experiment import rest_operations as experiment_rest_operations
 from .prompt import Prompt, PromptType
 from .prompt.client import PromptClient
 from .trace import migration as trace_migration
-from .optimization import Optimization
+from .optimization import OptimizationContextManager
 
 LOGGER = logging.getLogger(__name__)
 
 T = TypeVar("T")
-
-
-class OptimizationContextManager:
-    """
-    Context manager for handling optimization lifecycle.
-    Automatically updates optimization status to "completed" or "cancelled" based on context exit.
-    """
-
-    def __init__(
-        self,
-        client: "Opik",
-        dataset_name: str,
-        objective_name: str,
-        name: Optional[str] = None,
-        metadata: Optional[Dict[str, Any]] = None,
-    ):
-        """
-        Initialize the optimization context.
-
-        Args:
-            client: The Opik client instance
-            dataset_name: Name of the dataset for optimization
-            objective_name: Name of the optimization objective
-            name: Optional name for the optimization
-            metadata: Optional metadata for the optimization
-        """
-        self.client = client
-        self.dataset_name = dataset_name
-        self.objective_name = objective_name
-        self.name = name
-        self.metadata = metadata
-        self.optimization: Union[Optimization, None] = None
-
-    def __enter__(self) -> Union[str, None]:
-        """Create and return the optimization ID."""
-        try:
-            self.optimization = self.client.create_optimization(
-                dataset_name=self.dataset_name,
-                objective_name=self.objective_name,
-                name=self.name,
-                metadata=self.metadata,
-            )
-            if self.optimization:
-                return self.optimization.id
-            else:
-                return None
-        except Exception:
-            LOGGER.warning(
-                "Opik server does not support optimizations. Please upgrade opik."
-            )
-            LOGGER.warning("Continuing without Opik optimization tracking.")
-            return None
-
-    def __exit__(
-        self,
-        exc_type: Optional[Type[BaseException]],
-        exc_val: Optional[BaseException],
-        exc_tb: Optional[TracebackType],
-    ) -> Literal[False]:
-        """Update optimization status based on context exit."""
-        if self.optimization is None:
-            return False
-
-        try:
-            if exc_type is None:
-                self.optimization.update(status="completed")
-            else:
-                self.optimization.update(status="cancelled")
-        except Exception as e:
-            LOGGER.error(f"Failed to update optimization status: {e}")
-            return False
-        return False
 
 
 class Opik:
@@ -1205,7 +1130,7 @@ class Opik:
         _ = self._rest_client.optimizations.get_optimization_by_id(id)
         return optimization.Optimization(id=id, rest_client=self._rest_client)
 
-    def optimization_context(
+    def optimization_id(
         self,
         dataset_name: str,
         objective_name: str,
