+++ sdks/python/src/opik/api_objects/optimization/__init__.py
@@ -1,3 +1,3 @@
-from .optimization import Optimization
+from .optimization import Optimization, OptimizationContextManager
 
-__all__ = ["Optimization"]
+__all__ = ["Optimization", "OptimizationContextManager"]
